Moved the System Prompt creation to the chat provider.
