+++ java/src/main/java/com/datasqrl/ai/models/bedrock/BedrockChatSession.java
@@ -13,18 +13,7 @@ public class BedrockChatSession extends AbstractChatSession<BedrockChatMessage, 
   BedrockChatModel chatModel;
   BedrockTokenCounter tokenCounter;
 
-  private final String FUNCTION_CALLING_PROMPT = "To call a function, respond only with a JSON object of the following format: "
-      + "{\"function\": \"$FUNCTION_NAME\","
-      + "  \"parameters\": {"
-      + "  \"$PARAMETER_NAME1\": \"$PARAMETER_VALUE1\","
-      + "  \"$PARAMETER_NAME2\": \"$PARAMETER_VALUE2\","
-      + "  ..."
-      + "}. "
-      + "You will get a function response from the user in the following format: "
-      + "{\"function_response\" : \"$FUNCTION_NAME\","
-      + "  \"data\": { \"$RESULT_TYPE\": [$RESULTS] }"
-      + "}\n"
-      + "Here are the functions you can use:";
+
 
   public BedrockChatSession(BedrockChatModel model,
                             BedrockChatMessage systemMessage,
@@ -33,7 +22,7 @@ public class BedrockChatSession extends AbstractChatSession<BedrockChatMessage, 
     super(backend, sessionContext, null);
     this.chatModel = model;
     this.tokenCounter = BedrockTokenCounter.of(model);
-    this.systemMessage = convertMessage(combineSystemPromptAndFunctions(systemMessage.getTextContent()));
+    this.systemMessage = convertMessage(systemMessage);
   }
 
   @Override
@@ -95,28 +84,6 @@ public class BedrockChatSession extends AbstractChatSession<BedrockChatMessage, 
         .timestamp(Instant.now().toString())
         .numTokens(tokenCounter.countTokens(msg))
         .build();
-  }
-
-  private BedrockChatMessage combineSystemPromptAndFunctions(String systemPrompt) {
-    ObjectMapper objectMapper = new ObjectMapper();
-//    Note: This approach does not take into account the context window for the system prompt
-    String functionText = this.FUNCTION_CALLING_PROMPT
-        + this.backend.getFunctions().values().stream()
-        .map(RuntimeFunctionDefinition::getChatFunction)
-        .map(f ->
-            objectMapper.createObjectNode()
-                .put("type", "function")
-                .set("function", objectMapper.valueToTree(f))
-        )
-        .map(value -> {
-          try {
-            return objectMapper.writeValueAsString(value);
-          } catch (JsonProcessingException e) {
-            throw new RuntimeException(e);
-          }
-        })
-        .collect(Collectors.joining("\n"));
-    return new BedrockChatMessage(BedrockChatRole.SYSTEM, systemPrompt + "\n" + functionText + "\n", "");
   }
 
   private static String functionCall2String(BedrockFunctionCall fctCall) {
